requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
google-genai
python-dotenv
